fix: error while linting before building project #222
The lint job fails in github action because the storybook project depends on the built version of angular-ui.
When the lint command is executed before the build command, nodejs won't find the correct imports for the
storybook project.
In fact, the types to import are in the  folder generated by the build command. That's why the error occured.
To fix it, we either need to build the project before linting it or we change the import path to use the source code instead of the build version of the lib.

Signed-off-by: Axel SALLY <mrsallsa@protonmail.com>